// src/features/core/instructors/interfaces/instructors.interface.ts
export interface IInstructor {
    uuid: string;
    name: string;
    lastName: string;
    grado: string;
    profession: string;
    description: string;
    email: string;
    phone: string;
    address: string;
    yearsOfExperience: number;
    specialization: string;
    certifications: string[];
    education: string;
    languages: string[];
  }